Reorganized project
- Changed file structure
- Added software versionning, currently on version:  0.8.4-a
- Changed package.json app name to "discodes"
Go see on the docs website for more infos on versionning
