+++ app/src/lib/components/MiscComponents/AuthCheck.svelte
