+++ app/src/lib/components/NavBarComponents/NavBar.svelte
