+++ app/src/routes/auth/logout/+page.svelte
